replace catch by if check

+++ src/Lib/Session.php
@@ -137,14 +137,12 @@ class Session
 
 	public function logout()
 	{
-		$this->mem->logout($this->id());
-		try {
+		if ($this->initialized) {
+			$this->mem->logout($this->id());
 			$this->set('user', false);
 			fAuthorization::destroyUserInfo();
 			$this->setAuthLevel('guest');
 			$this->destroy();
-		} catch (Exception $e) {
-			// ignore if the user was already logged out
 		}
 	}
